+++ PowderGame/src/ch/client/gui/gui_component/GuiComp.cpp
@@ -5,13 +5,12 @@
 namespace ch {
 
 
-GuiComp::GuiComp(ch::EnumSides side, int xGap, int yGap, int width, int height) :
-	FittablePosition(side, xGap, yGap, width, height) {
+GuiComp::GuiComp(ch::EnumSides side, float xGap, float yGap, float width, float height)
+	: FittablePosition(side, xGap, yGap, width, height) {
 
 }
 
 GuiComp::~GuiComp() { }
-
 
 
 }
