+++ PowderGame/src/ch/client/gui/gui_component/GuiComp_Button.cpp
@@ -7,27 +7,27 @@
 namespace ch {
 
 
-GuiComp_Button::GuiComp_Button(ch::EnumSides side, int xGap, int yGap, int width, int height,
-                               EnumButtonKinds buttonKind, std::wstring text, int text_size,
+GuiComp_Button::GuiComp_Button(ch::EnumSides side, float xGap, float yGap, float width, float height,
+                               EnumButtonKinds buttonKind, std::wstring text, float text_size,
                                EnumButtonMove button_move_type)
   : GuiComp(side, xGap, yGap, width, height), text_size_(text_size), move_type_(button_move_type) {
 
   sf_text_.setFont(Game::getInstance()->getFont(ch::EnumGameFonts::BASIC));
   sf_text_.setString(text);
   sf_text_.setFillColor(sf::Color::Black);
-  sf_text_.setCharacterSize(text_size);
+  sf_text_.setCharacterSize(static_cast<unsigned int>(text_size));
 
   switch (buttonKind) {
 
   case EnumButtonKinds::COMMON:
   default:
-    Game::getInstance()->getTextureLoader().setTexture_Button(sf_sprite_[0], sf_sprite_[1], sf_sprite_[2], width, height);
+    Game::getInstance()->getTextureLoader().setTexture_Button(sf_sprite_[0], sf_sprite_[1], sf_sprite_[2], static_cast<int>(width), static_cast<int>(height));
   }
 }
 
 GuiComp_Button::~GuiComp_Button() { }
 
-void GuiComp_Button::setTransformedAABB_(int GUI_xPos, int GUI_yPos, int GUI_width_scaled, int GUI_height_scaled, float GUI_scale) {
+void GuiComp_Button::setTransformedAABB_(float GUI_xPos, float GUI_yPos, float GUI_width_scaled, float GUI_height_scaled, float GUI_scale) {
   sf_sprite_[0].setPosition(getTransformedAABB().pos);
   sf_sprite_[1].setPosition(getTransformedAABB().pos);
   sf_sprite_[2].setPosition(getTransformedAABB().pos);
